requirements.txt
gradio==3.50.2
google-cloud-language==2.11.1
requests==2.31.0
Pillow==10.0.0
openai==0.28.1